requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.13
botocore==1.24.13
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.2
django-allauth==0.49.0
django-countries==7.3.2
django-debug-toolbar==3.2.4
django-storages==1.12.3
idna==3.3
jmespath==0.10.0
oauthlib==3.2.0
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==4.1.1
urllib3==1.26.8
